requirements.txt
Django==2.2.17
djangorestframework==3.12.2
flake8==3.8.4
importlib-metadata==3.3.0
Markdown==3.3.3
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.4
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
